feat: add Timer Component
This component contains the following:

A timer label for session initializationn with the id 'timer-label' to the page to display the string 'Session' for a session initialization message.
Closes#7

A element with the id 'time-left' to display the time left for a session in mm:ss format, which updates correctly whether the session is running or paused.
Closes#8

A  button with the id  of 'start_stop'
Closes#9

A  button with the id  of 'reset'
Closes#10