Add audio functionality for countdown timer
This commit adds the necessary audio functionality for the countdown timer, as specified in user stories #28, #29, and #30. The "beep" audio element will now stop and rewind to the beginning when the "reset" button is clicked (#28), and the audio file is now required to be at least one second long (#29). Additionally, when the countdown reaches zero, the "beep" sound will play (#30) to indicate that time is up. This functionality was implemented using HTML5 audio tags with corresponding id attributes.

Fixes #28 #29 #30